add error to product services
