+++ pkg/microservice/aslan/core/environment/service/revision.go
@@ -298,6 +298,7 @@ func compareServicesRev(serviceTmplNames []string, services []*commonmodels.Prod
 				Deleted:     true,
 				Type:        service.Type,
 				Containers:  service.Containers,
+				Error:       service.Error,
 			}
 			serviceRevs = append(serviceRevs, serviceRev)
 		} else {
@@ -324,6 +325,7 @@ func compareServicesRev(serviceTmplNames []string, services []*commonmodels.Prod
 				ServiceName:     service.ServiceName,
 				Type:            service.Type,
 				CurrentRevision: service.Revision,
+				Error:           service.Error,
 				NextRevision:    maxServiceTmpl.Revision,
 			}
