+++ pkg/microservice/aslan/core/environment/service/types.go
@@ -47,6 +47,7 @@ type SvcRevision struct {
 	NextRevision    int64                     `json:"next_revision"`
 	Updatable       bool                      `json:"updatable"`
 	DeployStrategy  string                    `json:"deploy_strategy"`
+	Error           string                    `json:"error"`
 	Deleted         bool                      `json:"deleted"`
 	New             bool                      `json:"new"`
 	Containers      []*commonmodels.Container `json:"containers,omitempty"`
